refactor(cli): extract maptool subcommand to internal/maptoolcli
