+++ internal/maptoolcli/format.go
@@ -1,4 +1,4 @@
-package main
+package maptoolcli
 
 import (
 	"encoding/json"
