+++ internal/maptoolcli/render_test.go
@@ -1,4 +1,4 @@
-package main
+package maptoolcli
 
 import (
 	"context"
